+++ src/mod/Stats/Command/RegisterCommand.cpp
@@ -23,13 +23,13 @@ using namespace ll::i18n_literals;
 
 namespace stats::command {
 struct StatsGui {
-    StatsType   StatsType;
+    StatsType   category;
     std::string playerName;
 };
 
 struct StatsRank {
-    StatsType   StatsType;
-    std::string type;
+    StatsType   category;
+    std::string stat;
 };
 
 void registerCommand() {
@@ -42,16 +42,16 @@ void registerCommand() {
                     .alias("command.stats.alias"_tr());
     cmd.overload<StatsGui>()
         .text("gui")
-        .optional("StatsType")
+        .optional("category")
         .execute([&](CommandOrigin const& origin, CommandOutput& output, StatsGui const& param) {
             auto* entity = origin.getEntity();
             if (entity == nullptr || !entity->isType(::ActorType::Player)) {
-                return output.error("command.error.notplayer"_tr());
+                return output.error("command.error.not_player"_tr());
             }
             auto* player = static_cast<Player*>(entity);
-            // lk::MyMod::getInstance().getSelf().getLogger().info("cmd {} {}", player->getRealName(), param.statsType);
-            if (isValidStatsType(param.StatsType)) {
-                form::sendStatsGui(*player, param.StatsType);
+            // lk::MyMod::getInstance().getSelf().getLogger().info("cmd {} {}", player->getRealName(), param.category);
+            if (isValidStatsType(param.category)) {
+                form::sendStatsGui(*player, param.category);
             } else {
                 form::sendMainGui(*player);
             }
@@ -60,43 +60,43 @@ void registerCommand() {
     cmd.overload<StatsGui>()
         .text("player")
         .required("playerName")
-        .required("StatsType")
+        .required("category")
         .execute([&](CommandOrigin const& origin, CommandOutput& output, StatsGui const& param) {
             auto* entity = origin.getEntity();
             if (entity == nullptr || !entity->isType(::ActorType::Player)) {
-                return output.error("command.error.notplayer"_tr());
+                return output.error("command.error.not_player"_tr());
             }
             auto* player = static_cast<Player*>(entity);
             if (param.playerName.empty()) {
                 return output.error("command.error.player_name_empty"_tr());
             }
             auto const* cached = findCachedStatsByName(param.playerName);
-            if (!cached) return output.error("command.error.find_player"_tr());
+            if (!cached) return output.error("command.error.player_not_found"_tr());
             form::sendStatsGui(
                 *player,
                 mce::UUID(cached->first.uuid),
                 param.playerName,
-                param.StatsType
+                param.category
             );
         });
 
     cmd.overload<StatsRank>()
         .text("rank")
-        .required("StatsType")
-        .optional("type")
+        .required("category")
+        .optional("stat")
         .execute([&](CommandOrigin const& origin, CommandOutput& output, StatsRank const& param) {
             auto* entity = origin.getEntity();
             if (entity == nullptr || !entity->isType(::ActorType::Player)) {
-                return output.error("command.error.notplayer"_tr());
+                return output.error("command.error.not_player"_tr());
             }
             auto* player = static_cast<Player*>(entity);
-            if (!isValidStatsType(param.StatsType)) {
-                return output.error("command.error.invalid_stats_type"_tr());
+            if (!isValidStatsType(param.category)) {
+                return output.error("command.error.invalid_stats_category"_tr());
             }
-            if (param.StatsType == StatsType::custom && param.type.empty()) {
-                return output.error("command.error.rank_type_required"_tr());
+            if (param.category == StatsType::custom && param.stat.empty()) {
+                return output.error("command.error.rank_stat_required"_tr());
             }
-            form::sendRankGui(*player, param.StatsType, param.type);
+            form::sendRankGui(*player, param.category, param.stat);
         });
 }
 } // namespace stats::command